// src/types/fictioneers-api.ts
export interface components {
  schemas: {
// ...
    CreateUserDeserializer: {
      /**
       * Timeline ID the user should be placed on
       * @example 5UNDJoOUBDfSoMlW97a
       */
      published_timeline_id: string;
      /**
       * Timezone for the narrative story progression system
       * @example Europe/London
       */
      timezone: string;
      /**
       * Disable Time Guards
       * @description Feature flag to bypass time based guards per user.
       * @default false
       * @example false
       */
      disable_time_guards?: boolean;
      /**
       * Pause At Beats
       * @description Feature flag to toggle pausing execution of progression at beat markers.
       * @default false
       * @example false
       */
      pause_at_beats?: boolean;
    };
// ...
    EphemeralTokenSerializer: {
      /**
       * Primary Key
       * @description Access Token used to authenticate with Audience APIs.
       * @example ryJhbGciOiJSUzI1NiIsImtpZCI6IjQ7OTQ5ZDdkNDA3ZmVjOXIyYWM4ZDYzNWVjYmEwYjdhOTE0LWQ4ZmIiLCJ0eXAiOiJK
       */
      access_token: string;
      /**
       * Expires In
       * @description Time in seconds until the ID Token expires.
       * @example 3600
       */
      expires_in: number;
    };
// ...
    ProgressEventsDeserializer: {
      /**
       * Max Steps
       * @description The maximum amount of steps that should be progressed
       * @example 1
       */
      max_steps?: number;
      /**
       * Pause At Beats
       * @description Should beat guards be respected when progressing
       * @example true
       */
      pause_at_beats?: boolean;
    };
    /**
     * ResponseListSerializer[UserTimelineEventSerializer]
     * @description Base serializer class for all list responses, inheriting error and metadata from
     * ResponseSerializer.
     */
    ResponseListSerializer_UserTimelineEventSerializer_: {
      /** Data */
      data?: components["schemas"]["UserTimelineEventSerializer"][];
      /** Error */
      error?: components["schemas"]["ErrorSerializer"] | null;
      /** Meta */
      meta?: components["schemas"]["MetaSerializer"] | null;
    };
    /**
     * ResponseListSerializer[UserTimelineEventStateChangeSerializer]
     * @description Base serializer class for all list responses, inheriting error and metadata from
     * ResponseSerializer.
     */
    ResponseListSerializer_UserTimelineEventStateChangeSerializer_: {
      /** Data */
      data?: components["schemas"]["UserTimelineEventStateChangeSerializer"][];
      /** Error */
      error?: components["schemas"]["ErrorSerializer"] | null;
      /** Meta */
      meta?: components["schemas"]["MetaSerializer"] | null;
    };
    /**
     * ResponseSerializer[UserSerializer]
     * @description Base serializer class for all responses.
     */
    ResponseSerializer_UserSerializer_: {
      data?: components["schemas"]["UserSerializer"];
      /** Error */
      error?: components["schemas"]["ErrorSerializer"] | null;
      /** Meta */
      meta?: components["schemas"]["MetaSerializer"] | null;
    };
    /**
     * ResponseSerializer[UserStoryStateSerializer]
     * @description Base serializer class for all responses.
     */
    ResponseSerializer_UserStoryStateSerializer_: {
      data?: components["schemas"]["UserStoryStateSerializer"];
      /** Error */
      error?: components["schemas"]["ErrorSerializer"] | null;
      /** Meta */
      meta?: components["schemas"]["MetaSerializer"] | null;
    };
// ...
    UserTimelineEventSerializer: {
      /**
       * Id
       * @example kNSI3QNGVj3OwByy43ze
       */
      id: string;
      /**
       * Thread Id
       * @example PYztAWXDPnz4BM6qDzx4
       */
      thread_id?: string;
      /**
       * Row Id
       * @example 7D83C0It0Kpm62ZQ1gt9
       */
      row_id: string;
      /**
       * Previous In Row Timeline Event Id
       * @example IM50smQ3suY3nGOXNjp4
       */
      previous_in_row_timeline_event_id?: string;
      /**
       * Next In Row Timeline Event Id
       * @example s1sad1DkWZephtcJDmtM
       */
      next_in_row_timeline_event_id?: string;
      /**
       * Related Timeline Event Ids
       * @example [
       *   "MN7S90NBDJNSJS"
       * ]
       */
      related_timeline_event_ids: string[];
      /**
       * Available Step Index
       * @description Identifies the step index in which this event becomes available.
       * @example 1
       */
      available_step_index: number;
      /**
       * Available Beat
       * @description Identifies the beat in which this event becomes available.
       */
      available_beat?: components["schemas"]["BeatSerializer"];
      state: components["schemas"]["UserTimelineEventState"];
      /**
       * @description Future states which can be mutated by the user over API.
       * @example [
       *   "STARTED",
       *   "SKIPPED",
       *   "COMPLETED"
       * ]
       */
      available_states: components["schemas"]["UserInvokedTimelineEventState"][];
      /**
       * Narrative Event Id
       * @example 0nfWvgIlDZVk5yndCLKm
       */
      narrative_event_id: string;
      narrative_event_type: components["schemas"]["NarrativeEventType"];
      /**
       * Narrative Event Title
       * @example My narrative event
       */
      narrative_event_title?: string;
      /**
       * Narrative Event Description
       * @example My narrative event description
       */
      narrative_event_description?: string;
      /**
       * Narrative Event Content
       * @example [
       *   {
       *     "content_id": "MN7S90NBDJ",
       *     "content_type": "BlogPost",
       *     "provider_id": "Contentful"
       *   },
       *   {
       *     "content_id": "EL4S90NNDL",
       *     "content_type": "BlogPost",
       *     "provider_id": "Contentful"
       *   }
       * ]
       */
      narrative_event_content: components["schemas"]["ContentIntegrationSerializer"][];
      /**
       * Narrative Event State Changes Content
       * @description State change content associated with this event.
       * @example {
       *   "AVAILABLE": [
       *     {
       *       "content_id": "MN7S90NBDJ",
       *       "content_type": "BlogPost",
       *       "provider_id": "Contentful"
       *     },
       *     {
       *       "content_id": "EL4S90NNDL",
       *       "content_type": "BlogPost",
       *       "provider_id": "Contentful"
       *     }
       *   ],
       *   "STARTED": [
       *     {
       *       "content_id": "MN7S90NBDJ",
       *       "content_type": "BlogPost",
       *       "provider_id": "Contentful"
       *     }
       *   ]
       * }
       */
      narrative_event_state_changes_content: {
        [key: string]: components["schemas"]["ContentIntegrationSerializer"][];
      };
    };
// ...
  };
}